logging via Hydra and MPSpretraining.
