+++ configs/config.yaml
@@ -5,3 +5,11 @@ defaults:
   - model/dis: test
   - pretrain/mps: test
   - pretrain/dis: test
+  - override hydra/job_logging: debug
+  - override hydra/hydra_logging: default
+
+  # to turn off hydra logging: set the overrides above to none and add https://stackoverflow.com/questions/65104134/disable-file-output-of-hydra?rq=3:
+  # hydra:
+  #   output_subdir: null
+  #   run: 
+  #     dir: .
