+++ configs/pretrain/mps/test.yaml
@@ -13,7 +13,7 @@ auto_stack: True
 auto_unbind: False
 # cls_pos should be inherited from the mps config file
 title: "${dataset.title}"
-goal_acc: None
+goal_acc: null
 print_early_stop: True
 print_updates: True
 batch_size: 32 # samples loaded per categorisation step for all classes involved
