+++ src/sampling.py
@@ -1,6 +1,7 @@
 # script with sampling functions
 import torch
-
+import logging
+logger = logging.getLogger(__name__)
 
 #---------------------------------------------------------------------------------------------------------------------
 #---------------------------------------------------------------------------------------------------------------------
